refactor(http): migrate server/session/gateway config to http-types (#852)
Move the remaining pure McpHttpConfig sub-configs out of
dcc-mcp-http::config and into dcc-mcp-http-types::config:

- ServerConfig: core listener identity and transport knobs
- SessionConfig: session TTL and tool-list cache switch
- GatewayConfig: gateway election, routing, discovery, and admin UI knobs

These are all value types with no axum/tokio/reqwest/pyo3 dependency.
Only the McpHttpConfig aggregate remains in dcc-mcp-http, preserving the
application-level composition point while letting external Rust tooling
consume the complete sub-config contract through dcc-mcp-http-types.

Compatibility is preserved by re-exporting the migrated types from the
historical dcc-mcp-http::config path. The migrated structs derive serde
and use serde(default) so partial/minimal config bodies inherit the same
documented defaults as before.

Verification:
- cargo check --workspace
- cargo check -p dcc-mcp-http-types -p dcc-mcp-http
- cargo clippy -p dcc-mcp-http-types -p dcc-mcp-http --all-targets -- -D warnings
- cargo test -p dcc-mcp-http-types -p dcc-mcp-http --lib
