+++ crates/dcc-mcp-http-types/src/lib.rs
@@ -25,7 +25,7 @@
 //! |-------------|-----------------------------------------------------------|
 //! | crate root  | [`TruncationEnvelope`], [`SseChunkFrame`] + chunk helpers |
 //! | [`error`]   | [`HttpError`] / [`HttpResult`] error taxonomy             |
-//! | [`config`]  | [`ServerSpawnMode`], [`JobRecoveryPolicy`], [`JobConfig`], [`WorkflowConfig`], [`TelemetryConfig`], [`FeatureFlags`], [`InstanceConfig`], [`QueueConfig`] |
+//! | [`config`]  | [`ServerConfig`], [`SessionConfig`], [`GatewayConfig`], [`ServerSpawnMode`], [`JobRecoveryPolicy`], [`JobConfig`], [`WorkflowConfig`], [`TelemetryConfig`], [`FeatureFlags`], [`InstanceConfig`], [`QueueConfig`] |
 //! | [`output`]  | [`output::OutputStream`] and [`output::OutputEntry`] output capture wire types |
 //! | [`prompts`] | [`prompts::PromptSpec`], [`prompts::PromptsSpec`], and related prompt spec types |
 //! | [`resources`] | [`resources::ProducerContent`] / [`resources::ResourceError`] resource values |
@@ -38,10 +38,13 @@
 //! | Lives here now              | Stays in `dcc-mcp-http` for now |
 //! |-----------------------------|----------------------------------|
 //! | [`TruncationEnvelope`]      | `McpHttpConfig` aggregate       |
-//! | [`SseChunkFrame`]           | `ServerConfig` (uses `IpAddr`)  |
-//! | [`chunk_sse_data`]          | `SessionConfig`                  |
-//! | [`format_chunked_sse`]      | `GatewayConfig`                  |
+//! | [`SseChunkFrame`]           | `McpHttpConfig` aggregate       |
+//! | [`chunk_sse_data`]          |                                  |
+//! | [`format_chunked_sse`]      |                                  |
 //! | [`error::HttpError`]        |                                  |
+//! | [`config::ServerConfig`]    |                                  |
+//! | [`config::SessionConfig`]   |                                  |
+//! | [`config::GatewayConfig`]   |                                  |
 //! | [`config::ServerSpawnMode`] |                                  |
 //! | [`config::JobRecoveryPolicy`]|                                 |
 //! | [`config::JobConfig`]       |                                  |
